+++ spec/gush_spec.rb
@@ -4,7 +4,7 @@ describe Gush do
   describe ".gushfile" do
     let(:path) { Pathname("/tmp/Gushfile.rb") }
 
-    context "Gushfile.rb is missing from pwd" do
+    context "Gushfile is missing from pwd" do
       it "raises an exception" do
         path.delete if path.exist?
         Gush.configuration.gushfile = path
@@ -13,7 +13,7 @@ describe Gush do
       end
     end
 
-    context "Gushfile.rb exists" do
+    context "Gushfile exists" do
       it "returns Pathname to it" do
         FileUtils.touch(path)
         Gush.configuration.gushfile = path